Setting up unit tests

+++ app/src/main/java/com/bojio/mugger/listings/ListingUtils.java
@@ -27,8 +27,8 @@ public class ListingUtils {
     return getDayTimestamp(timestamp1) == getDayTimestamp(timestamp2);
   }
 
-  public static int daysApart(long timestamp1, long timestamp2) {
-    return (int) ((ListingUtils.getDayTimestamp(timestamp1) - ListingUtils
+  public static long daysApart(long timestamp1, long timestamp2) {
+    return ((ListingUtils.getDayTimestamp(timestamp1) - ListingUtils
         .getDayTimestamp(timestamp2)) / (1000 * 60 * 60 * 24)); // 1 week
   }
